fix new month formatting, use goals tree select
