+++ src/views/HomeView.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { useTaskStore } from '../stores/TaskStore'
-import GoalSelect from '../components/GoalTreeSelect.vue'
+import GoalSelect from '../components/GoalSelect.vue'
 import WeekTimelineEntry from '../components/WeekTimelineEntry.vue'
 
 const store = useTaskStore()
